// src/api/model/depotTransaction.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DepotTransactionDocumentType } from './depotTransactionDocumentType';
import type { Depot } from './depot';
import type { ProductPlan } from './productPlan';
import type { EmployeeExtendedRef } from './employeeExtendedRef';
import type { DepotRowItem } from './depotRowItem';

/**
 * Depo hareketi DTO'su
 */
export interface DepotTransaction {
  /** ID */
  id?: number;
  /** FiÅŸ tarihi */
  transactionDate: string;
  /** FiÅŸ numarasÄ± */
  documentNo: string;
  /** FiÅŸ tÃ¼rÃ¼ */
  documentType: DepotTransactionDocumentType;
  depot: Depot;
  productPlan: ProductPlan;
  employee: EmployeeExtendedRef;
  /** Hareket kalemleri listesi */
  rowItems: DepotRowItem[];
}