// src/api/model/depotRowItem.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MaterialCard } from './materialCard';

/**
 * Hareket kalemleri listesi
 */
export interface DepotRowItem {
  /** ID */
  id?: number;
  /** Ä°ÅŸlem tarihi */
  transactionDate: string;
  material: MaterialCard;
  /** Miktar */
  quantity: number;
}