// src/api/model/productPlan.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PlanStatusEnum } from './planStatusEnum';
import type { EmployeeExtendedRef } from './employeeExtendedRef';
import type { ProductionPlanProduct } from './productionPlanProduct';

/**
 * Malzeme talebinin iliÅŸkili olduÄŸu Ã¼retim planÄ±
 */
export interface ProductPlan {
  /** @nullable */
  id?: number | null;
  code: string;
  description: string;
  planStatus: PlanStatusEnum;
  employee: EmployeeExtendedRef;
  endTime: string;
  prodPlans?: ProductionPlanProduct[];
}