// src/api/model/employeeExtendedRef.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Branch } from './branch';
import type { Department } from './department';

export interface EmployeeExtendedRef {
  /** @nullable */
  id?: number | null;
  name?: string;
  surname?: string;
  companyBranch?: Branch;
  department?: Department;
  email: string;
}